v. 1.0.2

+++ updateOrDeleteFoodItem.py
@@ -7,27 +7,6 @@ def display_food_items():
             print(f"{i}. {item['name']}")
         return food_items
 
-def main():
-    food_items = display_food_items()
-    choice = input("Enter the number of the food item to update/delete (or 0 to go back): ")
-
-    if choice.isdigit() and 0 <= int(choice) <= len(food_items):
-        choice = int(choice)
-        if choice == 0:
-            return
-        
-        choice -= 1  # Adjust for zero-based indexing
-
-        print("0. Cancel and Return to Main Page")
-        print("1. Update Food Item")
-        print("2. Delete Food Item")
-        action = input("Enter a number to represent an action: ")
-
-        if action == '1':
-            update_food_item(food_items, choice)
-        elif action == '2':
-            delete_food_item(food_items, choice)
-
 def update_food_item(food_items, index):
     print("Updating food item:", food_items[index]['name'])
     print("Leave blank to keep current value.")
@@ -49,12 +28,41 @@ def update_food_item(food_items, index):
     print("Food item updated successfully.")
 
 def delete_food_item(food_items, index):
-    del food_items[index]
+    "deletes a registered food item "
+    food_item_name = food_items[index]['name']
+    confirmation = input(f"Are you sure you want to permanently delete {food_item_name}? (y/n): ").strip().lower()
+    
+    if confirmation == 'y':
+        del food_items[index]
 
-    with open('food_items.json', 'w') as file:
-        json.dump(food_items, file, indent=4)
+        with open('food_items.json', 'w') as file:
+            json.dump(food_items, file, indent=4)
 
-    print("Food item deleted successfully.")
+        print(f"Food item {food_item_name} deleted successfully.")
+    else:
+        print(f"Deletion of {food_item_name} canceled.")
+
+def main():
+    food_items = display_food_items()
+    choice = input("Enter the number of the food item to update/delete (or 0 to go back): ")
+
+    # display choices and prompt user for choice.
+    if choice.isdigit() and 0 <= int(choice) <= len(food_items):
+        choice = int(choice)
+        if choice == 0:
+            return
+        
+        choice -= 1  # Adjust for zero-based indexing
+
+        print("0. Cancel and Return to Main Page")
+        print("1. Update Food Item")
+        print("2. Delete Food Item")
+        action = input("Enter a number to represent an action: ")
+
+        if action == '1':
+            update_food_item(food_items, choice)
+        elif action == '2':
+            delete_food_item(food_items, choice)
 
 
 if __name__ == "__main__":
